+++ src/uiprotect/cli/base.py
@@ -36,6 +36,7 @@ def run(ctx: typer.Context, func: Coroutine[Any, Any, T]) -> T:
     async def callback() -> T:
         return_value = await func
         await ctx.obj.protect.close_session()
+        await ctx.obj.protect.close_public_api_session()
         return return_value
 
     try: